+++ webmacs/download_manager/__init__.py
@@ -14,18 +14,81 @@
 # along with webmacs.  If not, see <http://www.gnu.org/licenses/>.
 
 import os
+import re
 import json
 import shlex
 import logging
+import itertools
 
 from PyQt5.QtCore import QObject, pyqtSlot as Slot, pyqtSignal as Signal, \
     QProcess
 
 from PyQt5.QtWebEngineWidgets import QWebEngineDownloadItem
 
-from ..minibuffer.prompt import Prompt, FSModel, PromptTableModel
-from .. import current_minibuffer
-from .. import hooks
+from .prompts import (DlChooseActionPrompt, DlOpenActionPrompt, DlPrompt,
+                      OverwriteFilePrompt)
+from .. import current_minibuffer, hooks, variables
+
+
+default_download_dir = variables.define_variable(
+    "default-download-dir",
+    "Change the default download dir.",
+    "",
+    type=variables.String(),
+)
+
+TEMPORARY_DOWNLOAD_DIR = None
+keep_temporary_download_dir = variables.define_variable(
+    "keep-temporary-download-dir",
+    "If set to True, the download dir proposed will be the last used one.",
+    False,
+    type=variables.Bool(),
+)
+
+
+def get_user_download_dir():
+    """
+    Returns the directory the user wants to put its download in.
+
+    Return None if there is no specific directory.
+    """
+    if keep_temporary_download_dir.value and TEMPORARY_DOWNLOAD_DIR:
+        return TEMPORARY_DOWNLOAD_DIR
+    return default_download_dir.value or None
+
+
+def extract_suggested_filename(path):
+    """
+    Split path from chromium into (dirname, filename).
+
+    Chromium always propose paths in ~/Download, so this method removes suffix
+    like (1) to get the "original" suggested filename.
+    """
+    path, fname = os.path.split(path)
+    fname = re.sub(r'\([0-9]+\)(?=\.|$)', "", fname)
+    return path, fname
+
+
+def find_unique_suggested_path(dirname, filename):
+    """
+    Do the same logic as chromium does to create a unique path suggestion.
+    """
+    fnames = set(os.listdir(dirname))
+
+    if filename not in fnames:
+        return os.path.join(dirname, filename)
+
+    parts = filename.split(".", 1)
+    if len(parts) == 1:
+        name, ext = parts[0], ""
+    else:
+        name, ext = parts[0], "." + parts[1]
+
+    counter = itertools.count(1)
+    while True:
+        newfname = "{}({}){}".format(name, next(counter), ext)
+        if newfname not in fnames:
+            return os.path.join(dirname, newfname)
 
 
 STATE_STR = {
@@ -39,67 +102,6 @@ STATE_STR = {
 
 def state_str(state):
     return STATE_STR.get(state, "Unknown state")
-
-
-class DlChooseActionPrompt(Prompt):
-    complete_options = {
-        "match": Prompt.FuzzyMatch,
-        "complete-empty": True,
-    }
-    value_return_index_data = True
-
-    def __init__(self, path, mimetype):
-        Prompt.__init__(self, None)
-        self.__actions = [
-            ("download", "download file on disk"),
-            ("open", "open file with external command"),
-        ]
-        name = os.path.basename(path)
-        if len(name) > 33:
-            name = name[:30] + "..."
-        self.label = "File {} [{}]: ".format(name, mimetype)
-
-    def completer_model(self):
-        return PromptTableModel(self.__actions)
-
-    def enable(self, minibuffer):
-        super().enable(minibuffer)
-        minibuffer.input().popup().selectRow(0)
-
-
-class DlOpenActionPrompt(Prompt):
-    complete_options = {
-        "match": Prompt.FuzzyMatch,
-        "complete-empty": True,
-    }
-    label = "Open file with:"
-    value_return_index_data = True
-
-    def __init__(self):
-        Prompt.__init__(self, None)
-
-    def completer_model(self):
-        return PromptTableModel([[e] for e in list_executables()])
-
-
-class DlPrompt(Prompt):
-    complete_options = {
-        "autocomplete": True
-    }
-
-    def __init__(self, path, mimetype):
-        Prompt.__init__(self, None)
-        self.label = "Download file [{}]:".format(mimetype)
-        self._dlpath = path
-
-    def completer_model(self):
-        # todo, not working
-        model = FSModel(self)
-        return model
-
-    def enable(self, minibuffer):
-        super().enable(minibuffer)
-        minibuffer.input().setText(self._dlpath)
 
 
 def download_to_json(dlitem):
@@ -192,13 +194,31 @@ class DownloadManager(QObject):
             self._start_download(dl)
 
         elif action == "download":
+            path = dl.path()
+            user_dir = get_user_download_dir()
+            if user_dir is not None:
+                _, name = extract_suggested_filename(path)
+                try:
+                    path = find_unique_suggested_path(user_dir, name)
+                except OSError as exc:
+                    logging.warning(
+                        "Can't use user_dir %s: %s", user_dir, str(exc)
+                    )
 
-            prompt = DlPrompt(dl.path(), dl.mimeType())
+            prompt = DlPrompt(path, dl.mimeType())
             path = minibuff.do_prompt(prompt)
             if path is None:
                 return
 
+            if os.path.isfile(path):
+                if not minibuff.do_prompt(OverwriteFilePrompt(path)):
+                    return
+
             dl.setPath(path)
+            if keep_temporary_download_dir.value:
+                global TEMPORARY_DOWNLOAD_DIR
+                TEMPORARY_DOWNLOAD_DIR = os.path.dirname(path)
+
             logging.info("Downloading %s...", dl.path())
 
             def finished():
@@ -230,24 +250,6 @@ class DownloadManager(QObject):
             os.unlink(path)
         except Exception:
             pass
-
-
-def list_executables():
-    try:
-        paths = os.environ["PATH"].split(os.pathsep)
-    except KeyError:
-        return []
-
-    executables = []
-    for path in paths:
-        try:
-            for file_ in os.listdir(path):
-                if os.access(os.path.join(path, file_), os.X_OK):
-                    executables.append(file_)
-        except Exception:
-            pass
-
-    return executables
 
 
 def get_shell():