+++ CHANGELOG.md
@@ -46,6 +46,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - **C-g** in webbuffer is now bound to **buffer-escape** instead of
   **buffer-unselect**, which does the same thing and send the Escape too (which
   closes popup and other things).
+- Added two variables, **default-download-dir** and
+  **keep-temporary-download-dir**.
 
 ### Changed